NuVe.jl Dokumentation

NuVe.jl Dokumentation

NuVe.jl Dokumentation

Benutzung

Funktionen

Typen

Index

Benutzung

Todo

Funktionen

NuVe.CalcFactors — Method.

CalcFactors(x::NamedArrays.NamedArray)

Berechnet die noch fehlenden Faktoren für die zu lösende Ungleichung.

NuVe.addUserConstraints — Method.

addUserConstraints(x::Array{VariableRef,1}, c::Array{Constraint,1})

Fügt die vom Nutzer festgelegten Randbedingungen dem Modell hinzu.

NuVe.decayCorrection — Method.

decayCorrection()

Diese Funktion gibt die zerfallskorrigierte Eingangsgröße der Proben eines gewählten Nuklidvektors im bei Settings angegeben Zeitraum zurück.

NuVe.df2namedarray — Method.

df2namedarray(x::DataFrame, rowname::String, columnname::String)

Diese Funktion wandelt einen DataFrame in ein Array{Float64, 2} um. Potentiell fehlende Werte werden durch 0 ersetzt.

NuVe.getInterval — Method.

getInterval()

Macht aus den Einstellungen (Settings) für Anfangs- und Endjahr ein Array, welches zusätzlich jedes Jahr dazwischen enthält.

NuVe.getSampleFromSource — Method.

getSampleFromSource(x::Symbol)

Ruft alle vorhanden Proben zu einem Nuklidvektor ab.

NuVe.nuclideParts — Method.

nuclideParts(x::NamedArrays.NamedArray)

Gibt ein NamedArray mit den Nuklidanteilen wieder. Die Summe aller Nuklide über jeder Probe ergibt 1.

NuVe.setObjectives — Method.

setObjectives(x::Array{VariableRef,1}, parts::NamedArrays.NamedArray, c::Array{Constraint,1})

Legt das Optimierungsziel des Modells aufgrund der Einstellungen (Settings) fest.

NuVe.df2array — Method.

df2array(x::DataFrame)

Wandelt ein DataFrame in ein Array{Any,2} um.

NuVe.diffDays — Method.

diffDays(sample_date::Array{Date,1})
diffDays(sample_date::Array{Date,1}, year::Int64)
diffDays(sample_date::Array, year::Array{Int64, 1})
diffDays(sample_date::Date)

Gibt die Anzahl an Tagen von der Probenahme von allen Proben eines Nuklidvektors und einem Referenzdatum an. Das Jahr und der Referenzmonat bzw. -tag werden aus den Einstellungen bezogen (Settings) oder das Jahr wird separat angegeben.

NuVe.getNuclideTypes — Method.

getNuclideTypes(s::String)

Gibt alle Nuklide eines bestimmten Strahlungstypes (α, β oder γ) zurück.

NuVe.getNuclidesFromConstraint — Method.

getNuclidesFromConstraint(x::Array{Constraint,1})

Gibt alle Nuklide zurück, welche in sich in einem Array von Constraints befinden.

NuVe.getSampleDate — Method.

getSampleDate(x::Symbol)

Ruft das Probenahmedatum bzw. Referenzdatum zu allen Proben eines Nuklidvektors ab.

NuVe.getSampleInfo — Method.

getSampleInfo(x::String, nv::Symbol)

Fragt zu einem gegebenen Nuklidvektor weitere Informationen ab (z. B. "date" oder "source" (Herkunftsort))

NuVe.getWeightsFromConstraint — Method.

getWeightsFromConstraint(x::Array{Constraint,1})

Gibt alle Wichtungen der Nuklide zurück, welche in sich in einem Array von Constraints befinden. Diese Funktion wird beispielsweise benötigt, wenn auf die Repräsentativität der Proben optimiert werden soll. Hierbei wird die Abweichung der Nuklide des Nuklidvektors gegenüber dem Mittelwert der Nuklide der Proben minimiert.

NuVe.nvdb — Method.

Datenbank mit allen Nuklidvektoren, Proben, Halbwertszeiten und Freigabewerten

NuVe.readDb — Method.

readDb(s::String, arg::Array{String,1})

Gibt die angegebene Tabelle mit den angegebenen Spalten als DataFrame zurück.

NuVe.readDb — Method.

readDb(s::String)

Gibt die angegebene Tabelle vollständig als DataFrame zurück.

NuVe.removeMissing — Method.

 removeMissing(x::DataFrame)

Ersetzt in einem DataFrame alle missing durch 0.

NuVe.travec — Method.

travec(x::Array)

Transponiert einen Vektor (n × 1) zu (1 × n).

Typen

NuVe.Constraint — Type.

Dieser Type beinhaltet die zusätzlichen Bedingungen zur Berechnung des Nuklidvektors. Dazu gehören das einzuschränkende Nuklid, das Relationszeichen, die Grenze und eine Wichtung.

NuVe.RefDate — Type.

Dieser Type beinhaltet die Angabe für den Referenztag und -monat für die Zerfallskorrektur zu einem bestimmten Jahr

NuVe.Settings — Type.

In diesem Type sind grundlegenden Einstellungen gespeichert. Dazu gehören der Name des Nuklidvektors, der Berechnungszeitraum, das Optimierungsziel, die zu berücksichtigen Messgeräte, ein möglicher Schwellenwert, das Referenzdatum RefDate (Tag und Monat) für die Zerfallskorrektur und die zu berücksichtigenden Freigabepfade für die jeweilligen Messverfahren.

Index

NuVe.Constraint

NuVe.RefDate

NuVe.Settings

NuVe.CalcFactors

NuVe.addUserConstraints

NuVe.decayCorrection

NuVe.df2array

NuVe.df2namedarray

NuVe.diffDays

NuVe.getInterval

NuVe.getNuclideTypes

NuVe.getNuclidesFromConstraint

NuVe.getSampleDate

NuVe.getSampleFromSource

NuVe.getSampleInfo

NuVe.getWeightsFromConstraint

NuVe.nuclideParts

NuVe.nvdb

NuVe.readDb

NuVe.readDb

NuVe.removeMissing

NuVe.setObjectives

NuVe.travec